refactor(cli): split soundpack_command.go by subcommand
The 1192-line file held five subcommand factories plus their helpers
all in one place — too dense for review and a magnet for future drift.
Split per subcommand:

  soundpack_command.go    newSoundpackCommand wiring only (~29 LOC)
  soundpack_init.go       init subcommand
  soundpack_list.go       list subcommand
  soundpack_validate.go   validate subcommand + validateResult
  soundpack_install.go    install subcommand + updateConfigForInstall
  soundpack_use.go        use subcommand
  soundpack_helpers.go    soundpackInfo, discover*, count*, copy*,
                          ExtractAllSoundKeys, detectPlatformFile,
                          categoryFromKey

soundpack_git.go still owns add/update/remove/status and reaches into
soundpack_helpers.go for the shared discovery types and helpers.

Pure mechanical move; no API change. Same package-level symbol set.

Resolves review finding #53.
